Create Team9_Final_Shiny_App_v420220620
Team 9 is building a hassle free descriptive and predictive football statistics prototype that Pluckers plans to add as a free service during the 2022 football season. Pluckers views this service as providing an edge to the fierce restaurant and bar competition during the football season. Team 9 will create a user interfaced shiny app allowing the decision maker to select a football team and the type of offensive play to generate various football statistics. The app will be uploaded with raw data on a week-to-week basis by the customer (Pluckers) to ensure their customers will have the latest information to enable fantasy football decisions. 

The team focused on providing the user with plentiful but useful descriptive data. The descriptive methodology includes items such as charts and tables that give the user information on the team and type of play selected. R Studio and Shiny allow the team to present side by side information ranging in charts and tables that otherwise would be cumbersome with other applications. In addition, R studio allows the team ‘sparce’ through large amounts of information to only present useful data for the user.  

References
•	https://growjo.com/company/Pluckers_Wing_Bar
•	https://www.kaggle.com/datasets/maxhorowitz/nflplaybyplay2009to2016